Version committed on 20211009, In this version, the structure is reformed to take the topological relation first and then to simulate

+++ NdgPhys/NdgMatSolver/NdgNonhydrostaticSolver/@NdgQuadratureFreeNonhydrostaticSolver3d/private/SWENonhydrostatic3d.h
@@ -46,9 +46,9 @@ void FetchFacialData(double *, double *, double *, int );
 
 void FetchDataInSparseMatrix(double *, double *, int , int );
 
-void FindUniqueElementAndSortOrder(double *, double *, int *, int, int );
+void FindUniqueElementAndSortOrder(double *, double *, double *, double, int );
 
-void FindFaceAndDirectionVector(double *, int *, int *, int *, int *, int , int , double *, double *, double *, int , int );
+void FindFaceAndDirectionVector(double *, double *, double *, double *, double *, int , int , double *, double *, double *, int , int );
 
 void FindGlobalBottomEdgeFace(int *, double *, int , int , int );
